src/app/interceptors/http-header.interceptor.ts
import { Injectable } from '@angular/core';
import {
  HttpRequest,
  HttpHandler,
  HttpEvent,
  HttpInterceptor
} from '@angular/common/http';
import { Observable } from 'rxjs';
@Injectable()
export class HeaderInterceptor implements HttpInterceptor {
  constructor() {}
  intercept(request: HttpRequest<any>, next: HttpHandler): Observable<HttpEvent<any>> {
    
    request = request.clone({
      setHeaders: {
        'x-rapidapi-host':  'skyscanner-skyscanner-flight-search-v1.p.rapidapi.com',
        'x-rapidapi-key': '537d031308msha6a42e9cfe0fa9bp1cddbejsn9c0c36619539',
        'Content-Type': 'application/x-www-form-urlencoded',
      }
    });
    return next.handle(request);
  }
}